PDFBOX-2576: fix cast in int multiplication to long
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1651952 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/util/DateConverter.java
@@ -431,7 +431,7 @@ public final class DateConverter
             // we parsed a time zone in default format
             int hrSign = (sign == '-' ? -1 : 1);
             tz.setRawOffset(restrainTZoffset(hrSign * (tzHours * MILLIS_PER_HOUR + tzMin *
-                                                       MILLIS_PER_MINUTE)));
+                                                       (long) MILLIS_PER_MINUTE)));
             tz.setID("unknown");
         }
         else if ( ! hadGMT)
